Prevent -k / -n when combined with -R from leaking into future tests.
The problem was that the -k / -n parameters would get sent to the
server for a reverse tests, but never cleared for future reverse tests,
because these parameters were only sent from the client if they were
non-zero.
.
This is fixed in two ways as a belt-and-suspenders approach.

1. The server always clears the variables used for -k / -n when
receiving parameters, so if the client doesn't send them, they
get cleared (set to zero) anyway.

2. The client sends values for -k / -n even if they're zero. Sending
zero values doesn't hurt anything (except for wasting a few bytes),
and it makes this use case work even for servers that don't have
the fix in item 1 above.

Fixes #1363.