Print the shorter usage hint on argument error
When an argument error is printed it is often the result of a simple
typo. A user familiar with iperf3 does not need the whole help text to
figure out how to fix the command-line. A brief pointer to the erroneous
argument is sufficient.

Until now, with the *long* help text printed after an argument error, a
user often needs to scroll up to find the message about the erroneous
argument.

Another user not yet familiar with iperf3 will still be informed how to
access the long help text---or might already have it opened on the side,
anyway.