Fix build on CentOS 7 / x86_64.
The issue is that the default compiler on CentOS 7 (and presumably
RHEL 7) is an old GCC that's too old to support atomic types.
In this case we attempt to approximate an atomic 64-bit type with
a normal 64-bit integer, and warn at runtime if the system doesn't
support lock-free operations with this data type.

This has been compile-tested and lightly run-tested on CentOS 7
x86_64. It might not work well with ancient non-GCC compilers or
32-bit hosts.