Fix busy polling in startup phase
During the startup phase, the socket file descriptors (socket-fds) of
the streams are added to the select fd-sets. However, these file
descriptors are not read or written, causing the select operation to be
non-blocking and leading to busy polling.

To address this issue, a separate fd-set is introduced specifically for
the startup phase, which includes only the control socket.

In most scenarios, this results in a short period of increased CPU usage
during startup. However, in more uncommon situations, such as
simulations that utilize target code, this creates unnecessary load on
the system.

Signed-off-by: Benjamin Beichler <benjamin.beichler@uni-rostock.de>