Implement fixes to make the control connection more robust.
These include various timeouts in Nread() to guarantee that it will
eventually exit, a 10-second timeout for each attempt to read data
from the network and an approximately 30-second overall timeout per
Nread() call.

Also the iperf3 server now checks the length of the received session
cookie, and errors out if this happens to be incorrect.

Reported by Jorge Sancho Larraz - Canonical.