json: add UNIX timestamp in milliseconds
This extends the "timestamp" field with a UNIX timestamp in milliseconds
representing the current local wall clock time.

The "timestamp" field now has a new subkey:

    "timestamp": {
      "time": "Tue, 25 Feb 2025 11:49:59 GMT",
      "timesecs": 1740484199,
      // new
      "timemillisecs": 1740484199926
    },

This is helpful when one does fine-grained network tests with iperf,
where iperf events need to be aligned with events coming from other
resources. A time resolution based on seconds is too coarse-grained
for that.

It is enough to use milliseconds here, as networking always comes
with small timing offsets and delays. Using microseconds wouldn't
add a benefit.

It is sufficient to only change the code at this place, as the
relative time offset for all further events already includes
milliseconds. Currently, iperf3 prints this time offset in seconds
encoded as a double with a microsecond granularity.

The parameter for the clock_gettime system call was changed from
CLOCK_MONOTONIC to CLOCK_REALTIME. Firstly, the first returns the
seconds since the system was booted (Linux) and secondly, this way
we align the behaviour with the implementation based on
gettimeofday().

Signed-off-by: Philipp Schuster <philipp.schuster@cyberus-technology.de>
On-behalf-of: SAP philipp.schuster@sap.com