iperf_auth: keep RSA_private_decrypt and RSA_public_encrypt errors
The returns values of RSA_public_encrypt and RSA_private_decrypt calls
are stored in unsigned types. These functions return -1 upon error.

These return values are checked after the call if they are negative, but
since they are stored in unsigned type, they are lost.

To avoid this, change the type of encryptedtext_len and plaintext_len to
a signed one.

Also, change some tab indents to space.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>