Resolve issue with openssl 3.5.3 encrypt (#1956)
* Set output buffer size prior to encrypt operation

When calling EVP_PKEY_encrypt with a non-null output buffer,
the output buffer length must be provided. Attempts to write
beyond this length will fail.

* Rename keysize to output_buffer_len

This more accurately represents the meaning; it is the minimum
buffer allocation necessary for an encrypt or decrypt operation
to succeed. This is the same size for both ciphertext and
cleartext, as padding is applied.

* Avoid out-of-bounds access when base64 decoding short strings

Check the length before reading memory.

* Don't over-allocate followed by partial reads

We know how much we expect to read; the input buffer
has a defined size. Allocate the exact buffer expected
instead of a larger one with a read expected to return
only partial data. This makes it simpler to follow the
logic and to avoid off-by-one errors.

* Add warnings on silent truncation

Input should not be this long, but makes the expectations
of the code clearer.