Decouple 'test_start' JSON generation in iperf_on_test_start
Replace the monolithic iperf_json_printf() call for generating the
'test_start' JSON object with direct cJSON_Add... calls.

This refactoring removes a significant maintenance overhead, as it
eliminates the need to rewrite the entire format string whenever a
new, particularly conditional, parameter (such as sock_prio) needs to
be added to the test parameters block.

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>